// src/models/Article.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type Article = {
    articleId?: string;
    title?: string;
    perex?: string;
    imageId?: string;
    readonly createdAt?: string;
    readonly lastUpdatedAt?: string;
};